newfrom_ai_python/pyproject.toml
[tool.poetry]
name = "main"
version = "0.1.0"
description = ""
authors = ["이규영 <lky4530@newfrom.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.27"
psycopg2 = "^2.9.9"
openai = "^1.12.0"
boto3 = "^1.34.49"
numpy = "^1.26.4"
python-multipart = "^0.0.9"
pypdf2 = "^3.0.1"
websockets = "^12.0"
redis = "^5.0.2"
itsdangerous = "^2.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
